updated helper.py in all test shared folder
