+++ sdk/communication/azure-communication-sms/tests/_shared/helper.py
@@ -40,3 +40,16 @@ class URIIdentityReplacer(RecordingProcessor):
         if 'url' in response:
             response['url'] = re.sub('/identities/([^/?]+)', '/identities/sanitized', response['url'])
         return response
+    
+class URIMsalUsernameReplacer(RecordingProcessor):
+    """Replace the MSAL username in request uri"""
+    def process_request(self, request):
+        resource = (urlparse(request.uri).netloc).split('.')[0]
+        request.uri = re.sub('common/userrealm/([^/.]+)', 'common/userrealm/sanitized@test', request.uri) 
+        request.uri = re.sub(resource, 'sanitized', request.uri)
+        return request
+    
+    def process_response(self, response):
+        if 'url' in response:
+            response['url'] = re.sub('common/userrealm/([^/.]+)', 'common/userrealm/sanitized@test', response['url'])
+        return response
